Use pincode as pk for SubDistrict.
Make an int field on SubDistrict called 'pincode' and use that
as primary key.  Also override the save method to populate this
field using district and pin.

This has the benefit that we can now fill in pincode in admin
instead of looking up id.  Also update admin.py to use the new
fields.

Also clean up SubDistrict.__str__, since that had to be changed
anyways to use pincode.